JavaProject/beautysalon/src/main/java/com/unibuc/beautysalon/service/ProductService.java
package com.unibuc.beautysalon.service;

import com.unibuc.beautysalon.entity.Product;
import com.unibuc.beautysalon.exception.ProductNotFoundException;
import com.unibuc.beautysalon.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.validation.annotation.Validated;

import java.util.List;
import java.util.Optional;

@Service
@Validated
public class ProductService {
    private final ProductRepository productRepository;
    @Autowired
    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }
    public List<Product> getAllProducts() {
        return productRepository.findAll();
    }

    public List<Product> getProductsByName(String name) {
        return productRepository.findByName(name);
    }

    public Product saveProduct(Product product) {
        return productRepository.save(product);
    }

    public void deleteProduct(Long productId) {
        productRepository.deleteById(productId);
    }
    public Product updateProductName(Long id, String newName) {
        Optional<Product> optionalProduct = productRepository.findById(id);
        if (optionalProduct.isPresent()) {
            Product product = optionalProduct.get();
            product.setName(newName);
            return productRepository.save(product);
        } else {
            throw new ProductNotFoundException(id);
        }
    }
}
